Fixed incompatibilies

+++ src/main/java/dev/ioannis/anemosparts/domain/PartSummaryDto.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -19,8 +20,8 @@ public class PartSummaryDto {
     private String description;
     private String oemNumber;
     private String partNumber;
-    private Double price;
-    private Integer quantity;
+    private BigDecimal price;
+    private Long quantity;
     private Optional<String> thumbnailSrc;
 
     private String brandIconSrc;
